js: split components into files and refactor code
