+++ invenio_damap/assets/semantic-ui/js/invenio_damap/index.js
@@ -1,4 +1,4 @@
-// This file is part of InvenioRDM
+// This file is part of Invenio-DAMAP
 // Copyright (C) 2023-2024 Graz University of Technology.
 // Copyright (C) 2023-2024 TU Wien.
 //
@@ -7,544 +7,8 @@
 
 import ReactDOM from "react-dom";
 import React from "react";
-import {
-  Button,
-  Checkbox,
-  Form,
-  Grid,
-  Icon,
-  Item,
-  Label,
-  Message,
-  Modal,
-  Popup,
-} from "semantic-ui-react";
-import { http } from "react-invenio-forms";
-
-import PropTypes from "prop-types";
-
-export class RadioGroupQuestion extends React.Component {
-  render() {
-    const { title, optionsAndValues, onChange, selectedValue } = this.props;
-    return (
-      <Form>
-        <Form.Group inline>
-          <label>{title}</label>
-
-          {Object.entries(optionsAndValues).map(([question, value]) => (
-            <Form.Radio
-              key={question + value}
-              label={question}
-              name={`radioGroup$title${title}`}
-              value={value}
-              checked={selectedValue === value}
-              onChange={() => {
-                onChange(value);
-              }}
-            />
-          ))}
-        </Form.Group>
-      </Form>
-    );
-  }
-}
-
-RadioGroupQuestion.propTypes = {
-  title: PropTypes.string.isRequired,
-  optionsAndValues: PropTypes.objectOf(PropTypes.any).isRequired,
-  onChange: PropTypes.func.isRequired,
-  selectedValue: PropTypes.any.isRequired,
-};
-
-export class UserQuestions extends React.Component {
-  question_types = ["personal_data", "sensitive_data"];
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      selectedValues: {
-        personal_data: "no",
-        sensitive_data: "no",
-      },
-    };
-  }
-
-  onChange = (key, value) => {
-    let newSelectedValues = { ...this.state.selectedValues };
-    newSelectedValues[key] = value;
-    this.setState({
-      selectedValues: newSelectedValues,
-    });
-    this.props.onChange(newSelectedValues);
-  };
-
-  render() {
-    return (
-      <div>
-        {this.question_types.map((question) => (
-          <RadioGroupQuestion
-            key={question}
-            title={`Does the dataset contain ${question}? *`.replace("_", " ")}
-            optionsAndValues={structuredClone({
-              Yes: "yes",
-              No: "no",
-            })}
-            selectedValue={this.state.selectedValues[question]}
-            onChange={(value) => {
-              this.onChange(question, value);
-            }}
-          ></RadioGroupQuestion>
-        ))}
-      </div>
-    );
-  }
-}
-UserQuestions.propTypes = {
-  onChange: PropTypes.func.isRequired,
-};
-
-// DMP Item begin
-
-export class DMPEntry extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      loading: false,
-    };
-  }
-
-  setLoading(loading) {
-    this.setState({ loading: loading });
-  }
-
-  render() {
-    const { dmp, record, onDmpSelected, checked } = this.props;
-
-    const alreadyAddedToDMP = this.props.dmp.datasets?.some((ds) => {
-      return (
-        ds.datasetId?.identifier === window.location.href ||
-        ds.datasetId?.identifier?.replace("uploads", "records") ===
-          window.location.href
-      );
-    });
-
-    return (
-      <Item>
-        <Item.Image size="mini">
-          <Checkbox
-            onChange={(e, data) => onDmpSelected(dmp, data.checked)}
-            checked={checked}
-          />
-        </Item.Image>
-        <Item.Content>
-          <Item.Header as="a">
-            {dmp.project?.title ?? "DMP ID: " + dmp.id}
-          </Item.Header>
-          <Item.Description>
-            {(dmp.project?.description ?? "").substring(0, 255)}
-          </Item.Description>
-          {alreadyAddedToDMP && (
-            <Item.Extra>
-              <Popup
-                content="If you link the record to this DMP, a new version will be created in DAMAP."
-                trigger={
-                  <Label icon="check" content="Already added" color="green" />
-                }
-              />
-            </Item.Extra>
-          )}
-        </Item.Content>
-      </Item>
-    );
-  }
-}
-
-DMPEntry.propTypes = {
-  record: PropTypes.object.isRequired,
-  dmp: PropTypes.object.isRequired,
-  onUpdate: PropTypes.func.isRequired,
-  onDmpSelected: PropTypes.func.isRequired,
-  checked: PropTypes.bool.isRequired,
-};
-
-// DMP Item end
-
-export class GenericMessage extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      message: props.message,
-    };
-  }
-
-  componentDidUpdate(prevProps) {
-    if (this.props.message.visible !== prevProps.message.visible) {
-      this.setState({
-        message: {
-          ...this.props.message,
-        },
-      });
-    }
-  }
-
-  handleDismiss = () => {
-    this.setState((prevState) => ({
-      message: {
-        ...prevState.message,
-        visible: false,
-      },
-    }));
-  };
-
-  render() {
-    const { icon, visible, type, header, content, errors } = this.state.message;
-    return (
-      <>
-        {visible && (
-          <Message
-            // We override the icon size because it's not controllable when used in a <Message>.
-            // Related bug: https://github.com/Semantic-Org/Semantic-UI/issues/6441
-            icon={<Icon name={icon} style={{ fontSize: "2em" }} />}
-            info={type === "info"}
-            warning={type === "warning"}
-            success={type === "success"}
-            error={type === "error"}
-            onDismiss={this.handleDismiss}
-            header={header}
-            content={content}
-            list={errors.map((error) => error.dmp.project.title)}
-          />
-        )}
-      </>
-    );
-  }
-}
-
-// Modal begin
-
-export class DMPModal extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      loading: props.loading,
-      dmps: [],
-      selectedDmps: [],
-      userQuestions: {},
-      message: {
-        visible: false,
-        icon: null,
-        type: null,
-        header: null,
-        content: null,
-        errors: [],
-      },
-    };
-  }
-
-  setLoading(loading) {
-    this.setState({ loading: loading });
-  }
-
-  onError(message) {
-    console.log("ERROR", message);
-    return null;
-  }
-
-  async fetchDMPs() {
-    this.setLoading(true);
-    try {
-      let dmpSearchResult = await http.get("/api/invenio_damap/damap/dmp");
-      this.setState({
-        dmps: dmpSearchResult.data.hits.hits,
-      });
-    } catch (error) {
-      this.setLoading(false);
-      console.error(error);
-    }
-    this.resetSelectedDmps();
-    this.setLoading(false);
-  }
-
-  componentDidMount() {
-    this.fetchDMPs();
-  }
-
-  onUserQuestionsChange = (questionsAndAnswers) => {
-    this.setState({
-      userQuestions: questionsAndAnswers,
-    });
-  };
-
-  onDmpSelected = (dmp, selected) => {
-    let x = this.state.selectedDmps.filter((d) => d !== dmp);
-    if (selected) {
-      x.push(dmp);
-    }
-
-    this.setState({
-      selectedDmps: x,
-    });
-  };
-
-  async onAddUpdateDataset(dmp_id, record) {
-    let { userQuestions } = this.state;
-    let body = userQuestions;
-
-    let response = await http.post(
-      `/api/invenio_damap/damap/dmp/${dmp_id}/dataset/${record.id}`,
-      body
-    );
-
-    return response;
-  }
-
-  showMessage(icon, type, header, content, errors) {
-    this.setState({
-      message: {
-        visible: true,
-        icon: icon,
-        type: type || "info",
-        header: header,
-        content: content,
-        errors: errors || [],
-      },
-    });
-  }
-
-  async addDatasetToDmps() {
-    this.setLoading(true);
-    this.resetMessage();
-    let { selectedDmps } = this.state;
-    let { record } = this.props;
-
-    let errors = [];
-    let responses = [];
-
-    for (let dmp of selectedDmps) {
-      responses.push(
-        this.onAddUpdateDataset(dmp.id, record).catch((e) => {
-          errors.push({ dmp, error: e });
-        })
-      );
-    }
-    await Promise.all(responses);
-
-    if (errors.length === 0) {
-      this.showMessage(
-        "check circle outline",
-        "success",
-        "Success!",
-        "Record was linked to DMP(s)."
-      );
-    } else if (errors.length === selectedDmps.length) {
-      this.showMessage(
-        "times circle outline",
-        "error",
-        "Error",
-        "Linking record to DMP(s) failed."
-      );
-    } else {
-      this.showMessage(
-        "warning sign",
-        "warning",
-        `Record was linked to DMP(s) with errors. Not linked/updated:`,
-        "",
-        errors
-      );
-    }
-    this.setLoading(false);
-  }
-
-  resetSelectedDmps = () => {
-    this.setState({
-      selectedDmps: [],
-    });
-  };
-
-  resetMessage() {
-    this.setState({
-      message: {
-        visible: false,
-        icon: null,
-        type: null,
-        header: null,
-        content: null,
-        errors: [],
-      },
-    });
-  }
-
-  handleModalClose = () => {
-    this.resetMessage();
-  };
-
-  render() {
-    const { open, handleClose, record } = this.props;
-    let { dmps, loading, selectedDmps } = this.state;
-
-    let buttonText = `Add or update dataset for ${selectedDmps.length} DMP(s)`;
-    let buttonIcon = "plus";
-    const isAddDMPButtonDisabled = selectedDmps.length === 0;
-
-    return (
-      <Modal
-        open={open}
-        onClose={handleClose}
-        onUnmount={this.handleModalClose}
-        className="share-modal"
-        role="dialog"
-        aria-labelledby="access-link-modal-header"
-        aria-modal="true"
-        tab-index="-1"
-      >
-        <Modal.Header id="dmp-modal-header">
-          <Icon name="share alternate" />
-          {"Link record to DMP"}
-        </Modal.Header>
-
-        <Modal.Content>
-          <Modal.Description>
-            <GenericMessage message={this.state.message}></GenericMessage>
-            <UserQuestions
-              onChange={this.onUserQuestionsChange}
-            ></UserQuestions>
-            <Item.Group divided>
-              {dmps.map((dmp, index) => (
-                <DMPEntry
-                  key={dmp.id}
-                  checked={selectedDmps.indexOf(dmp) > -1}
-                  onDmpSelected={this.onDmpSelected}
-                  floated="right"
-                  dmp={dmp}
-                  record={record}
-                  loading={loading}
-                  onUpdate={(event) => {
-                    this.fetchDMPs();
-                  }}
-                ></DMPEntry>
-              ))}
-            </Item.Group>
-          </Modal.Description>
-        </Modal.Content>
-
-        <Modal.Actions>
-          <Button
-            size="small"
-            floated="left"
-            onClick={(event) => {
-              this.fetchDMPs();
-            }}
-            icon
-            loading={loading}
-            labelPosition="left"
-          >
-            <Icon name="sync" />
-            {"Refresh DMPs"}
-          </Button>
-
-          <Button size="small" onClick={handleClose}>
-            {"Done"}
-          </Button>
-
-          <Button
-            primary
-            floated="right"
-            icon
-            loading={loading}
-            labelPosition="left"
-            onClick={async () => {
-              if (!isAddDMPButtonDisabled) {
-                await this.addDatasetToDmps();
-                this.resetSelectedDmps();
-                this.fetchDMPs();
-              }
-            }}
-            disabled={isAddDMPButtonDisabled}
-          >
-            <Icon name={buttonIcon} />
-            {buttonText}
-          </Button>
-        </Modal.Actions>
-      </Modal>
-    );
-  }
-}
-
-DMPModal.defaultProps = {
-  dmps: [],
-};
-
-DMPModal.propTypes = {
-  record: PropTypes.object.isRequired,
-  open: PropTypes.bool.isRequired,
-  handleClose: PropTypes.func.isRequired,
-};
-
-// Modal end
-
-// =====================================================
-
-// DMPButton begin
-
-export class DMPButton extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      record: props.record,
-      disabled: props.disabled,
-      open: props.open,
-    };
-  }
-
-  handleOpen = () => {
-    this.setState({
-      open: true,
-    });
-  };
-
-  handleClose = () => {
-    this.setState({
-      open: false,
-    });
-  };
-
-  render() {
-    const { disabled, open, record } = this.state;
-
-    return (
-      <>
-        <Button
-          fluid
-          onClick={this.handleOpen}
-          disabled={disabled}
-          primary
-          size="medium"
-          aria-haspopup="dialog"
-          icon
-          labelPosition="left"
-        >
-          <Icon name="plus square" />
-          {"Add to DMP"}
-        </Button>
-
-        <DMPModal open={open} handleClose={this.handleClose} record={record} />
-      </>
-    );
-  }
-}
-
-DMPButton.propTypes = {
-  disabled: PropTypes.bool,
-  record: PropTypes.object.isRequired,
-  open: PropTypes.bool,
-};
-
-DMPButton.defaultProps = {
-  disabled: false,
-  open: false,
-};
-
-// =========================================
+import { Grid } from "semantic-ui-react";
+import { DMPButton } from "./DMPButton";
 
 const element = document.getElementById("invenio-damap-render");
